feat(arazzo): GET /runners + indexed trigger gating; int32 version fields (P3c-2a)
Control-plane surface for the runner registry:
- New GET /runners endpoint (scope runs:read) — Runner/RunnerHostedVersion/RunnerPage
  schemas added to the OpenAPI doc and the server regenerated; the handler is a free
  re-wrap (Models.Runner.From(registration)) since Runner shares RunnerRegistration's shape.
- The catalog trigger (POST .../runs) now gates on a hosting runner: 409 "no-runner" when
  no registered runner hosts the targeted version with it loaded. IRunnerRegistry +
  MapArazzoControlPlane + the catalog handler take the registry.

Indexed hosting query (review feedback — the gate must not scan the whole registry):
- IRunnerRegistry gains IsVersionHostedAsync(baseWorkflowId, versionNumber); the gate calls it.
- Every backend answers it from an index, not a scan: SQL (Sqlite/Postgres/MySql/SqlServer)
  via a runner_hosted_versions child table indexed on (base, version), re-projected per
  register and cleaned on prune (FK cascade, or explicit delete on SQLite); Cosmos/Mongo via
  a loadedVersions projection queried with ARRAY_CONTAINS/$elemMatch; Redis via a per-version
  SET index; NATS via per-version index keys; Azure via a second index table; InMemory scans.
- Conformance gains 3 tests (hosted true/false, after-prune, after-reregistration) that
  exercise index maintenance.

int32 (review feedback — the (int)(long) double-cast signalled a missing format): versionNumber
and maxConcurrency now declare "format": "int32", so they generate as JsonInt32 with an
implicit int conversion; the double-casts are gone.

Verified: registry conformance (11 tests) green on InMemory + Sqlite + all 8 container backends
(podman); control-plane server tests 24/24; full slnx Debug build 0 errors.

+++ src/Corvus.Text.Json.Arazzo.Durability/Catalog/CatalogVersion.cs
@@ -26,7 +26,7 @@ public readonly partial struct CatalogVersion
     private static readonly JsonWriterOptions WriterOptions = new() { Indented = false, SkipValidation = true };
 
     /// <summary>Gets the minimal identity reference for this version.</summary>
-    public CatalogVersionRef Ref => new((string)this.BaseWorkflowId, (int)(long)this.VersionNumber, (string)this.WorkflowId);
+    public CatalogVersionRef Ref => new((string)this.BaseWorkflowId, this.VersionNumber, (string)this.WorkflowId);
 
     /// <summary>Gets the lifecycle status as the <see cref="CatalogStatus"/> enum.</summary>
     public CatalogStatus StatusValue => Enum.Parse<CatalogStatus>((string)this.Status);
